Fix errors and responsive
Soluciono errores de borrar fotos, y en el  front un poco de responsive
utilizacion del metodo clamp que es la puta hostia.

+++ Backend/src/main/java/com/prueba/api/authorized/UserController.java
@@ -65,7 +65,9 @@ public class UserController {
         }
         String fotoBorrar = userService.findFotoByUsername(username);
         response.put("OK", "Correcto");
-        userService.deleteByUsername(username);
+        if(!FileRemover.FOTO_DEFAULT.equals(fotoBorrar)){
+            userService.deleteByUsername(username);
+        }
         FileRemover.deleteFile(fotoBorrar);
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
